requirements.txt
fastapi==0.112.2
uvicorn==0.30.5

PyYAML~=6.0.3
pydantic[email]~=2.11.9
dotenv~=0.9.9
python-dotenv~=1.1.1
SQLAlchemy~=2.0.44
alembic~=1.17.0
asyncpg
greenlet
python-jose[cryptography]>=3.3.0
passlib~=1.7.4
python-multipart
